+++ src/app/(with-container)/(exercise)/exercise/thematic/[type]/page.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import ExercisePageInfoCard from "../../../panel/ExercisePageInfoCard";
 import { IPopularRoutine } from "@/types/IPopularRoutine";
 import VideoInfo from "../../panel/VideoInfo";
-import ReturnButton from "@/app/(with-footer)/my-center/members/panel/ReturnButton";
+import ReturnButton from "@/components/ReturnButton";
 import { useParams } from "next/navigation";
 import { thematicWorkoutCodesLabel, WorkoutKind } from "@/types/IRoutine";
